feat: Added charts using Charts library
Implemented new functions and components to generate charts
displaying expense data in the application. Utilized the Charts
library to create bar charts based on categories and their expenses.
Also implemented functions to process data and map categories to
appropriate colors for the charts.

Added the dependency on the Charts library in the build.gradle file
and included necessary imports in the code.

Modified components related to displaying charts and added support
for category colors in the chart bars.